+++ src/Manifolds.jl
@@ -469,6 +469,7 @@ include("manifolds/FlagStiefel.jl")
 include("manifolds/GeneralizedGrassmann.jl")
 include("manifolds/GeneralizedStiefel.jl")
 include("manifolds/HeisenbergMatrices.jl")
+include("manifolds/HermitianPositiveDefinite.jl")
 include("manifolds/Hyperbolic.jl")
 include("manifolds/Hyperrectangle.jl")
 include("manifolds/InvertibleMatrices.jl")
@@ -626,6 +627,7 @@ export Euclidean,
     Grassmann,
     HamiltonianMatrices,
     HeisenbergMatrices,
+    HermitianPositiveDefinite,
     Hyperbolic,
     Hyperrectangle,
     InvertibleMatrices,