+++ NEWS.md
@@ -5,6 +5,14 @@ All notable changes to ´Manifolds.jl´ will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.24] unreleased
+
+### Added
+
+* `HermitianPositiveDefinite` manifold of Hermitian positive definite matrices.
+* `MatrixSqrtManifoldPoint` as a generalization/replacement for the `SPDPoint`
+* `SymmetricPositiveDefinite` is now a `const` of `HermitianPositiveDefinite` over the reals
+
 ## [0.11.21] 2025-04-03
 
 ### Changed